TL: solved random collocation error with EquidistantSpline_Right
